Added typescript
Ran the npm install command with node version 14.17.0 and npm version 6.14.13.
Installed dependencies for using typescript in this project.
Updated the readme to include an explanation on node versions and a download link.
Updated frontend app files from js to tsx files.
Tsconfig file added as a result of initializing typescript in the repo.