updates
Modified git ignore to ignore the .env file that contains all of the secrets. Added back in the settings.py as it is a required file to run the application.
Added a urls.py under the app to link the views file and attach the function to return a response when the function is sent a GET request.
Updated the backend urls.py file to link the api/ endpoint with the app.urls file.
Added a pipenv to contain all of the packages used in this project.