#295: Receive and Store Biography
* Biography now saves to the database

Hello!

I've added a test case in "profiles_tests" that demonstrates the new functionality.

Whenever a user clicks "save" on their biography, the ProfileTextField's new prop, "handleSave" is called. This is passed down from the ProfileBiography component and is called on save.

I've also made it so that the ProfilePage component receives the 'accountID' prop, as it is necessary when making the post request to the server. As a result, the 'Profile' type has also been updated.

* Refactored biography view's POST method, adjusted to the new endpoint

Previously, the endpoint for biographies was "profiles/biography"
Now, it's "profiles/biography/{account_id}"

I've adjusted the post method and my test to reflect this.

Original PR: https://github.com/LetsMesh/Site/pull/293