Hotfix: Profile Tests Missing accountID Attribute
Fix for an earlier PR/Merge where I updated the "profile" type to have an accountID attribute.

The example profiles used in the front-end compilation were not updated, and thus would cause compilation/build failures.

I've added the attribute to the tests as a fix while I figure out how/if I will refactor the code.