handle button submit function
clicking submit button on /otp page will check the backend if the otp is correct.

TODO: currently user id is hard coded, which should be fetched according to the login page.
There could also be an extra styling for the otp page such as having a better alert for incorrect otp pin