Improve User experience for 2FA interaction
entering the OTP page now checks if the cookie required to validate OTP input exist. Else, user will be prompted to return to login page.

The session for OTP validation is now timed according to the OTP valid interval on backend. User will be prompted to return back to login page when the interval has passed.